cleaned up a lot of stuff
